pyproject.toml
[project]
name = "visual-grouding-qa-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "opencv-python>=4.12.0.88",
    "pytorch",
    "transformers",
]
